\name{keboola.redshift.r.client}
\alias{keboola.redshift.r.client.package}
\docType{package}
\title{
Amazon Redshift database driver
}
\description{
A set of methods useful for connecting to Amazon Redshift database. These functions are wrappers
around JDBC methods.
}
\details{
\tabular{ll}{
Package: \tab keboola.redshift.r.client\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-28\cr
License: \tab MIT\cr
}
}
\author{
Maintainer: Ondrej Popelka <ondrej.popelka@keboola.com>
}
\keyword{ database }
\examples{
~~ simple examples of the most important functions ~~
}


    