% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/client.R
\name{redshift.columnTypes}
\alias{redshift.columnTypes}
\title{Get list of columns in table and their datatypes}
\usage{
redshift.columnTypes(tableName)
}
\arguments{
\item{tableName}{Name of the table (including schema).}
}
\value{
Dataframe with columns 'column' and 'dataType'
}
\description{
Get list of columns in table and their datatypes
}

