% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/client.R
\docType{data}
\name{redshift.conn}
\alias{redshift.conn}
\title{Current database connection (to change, use redshift.connect())}
\format{\preformatted{ NULL
}}
\usage{
redshift.conn
}
\description{
Current database connection (to change, use redshift.connect())
}
\keyword{datasets}

