% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/client.R
\name{redshift.connect}
\alias{redshift.connect}
\title{Connect to Amazon Redshift database}
\usage{
redshift.connect(jdbcUrl, driver, username, password, schema)
}
\arguments{
\item{jdbcUrl}{JDBC connection string}

\item{driver}{Path to the JDBC driver}

\item{username}{Database user name}

\item{password}{Database password}

\item{schema}{Database schema}
}
\value{
void
}
\description{
Connect to Amazon Redshift database
}

