% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/client.R
\name{redshift.prepareStatement}
\alias{redshift.prepareStatement}
\title{Prepare a SQL query with quoted parameters}
\usage{
redshift.prepareStatement(sql, ...)
}
\arguments{
\item{sql}{SQL string, parameter placeholders are marked with ?}

\item{...}{Query parameters, number of parameters must be same as number of
question marks}
}
\value{
SQL string
}
\description{
Prepare a SQL query with quoted parameters
}

