% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/client.R
\name{redshift.saveDataFrame}
\alias{redshift.saveDataFrame}
\title{Save a dataframe to database using bulk inserts. The table will be created to accomodate to data frame columns.}
\usage{
redshift.saveDataFrame(dfRaw, table, rowNumbers = FALSE,
  incremental = FALSE, forcedColumnTypes)
}
\arguments{
\item{table}{Name of the table.}

\item{rowNumbers}{If true then the table will contain a column named 'row_num' with sequential row index}

\item{incremental}{If true then the table will not be recreated, only data will be inserted}

\item{forcedColumnTypes}{List of column names and their respective types in database.}

\item{df}{A data frame, column names of data frame must correspond to column names of table}
}
\value{
void
}
\description{
Save a dataframe to database using bulk inserts. The table will be created to accomodate to data frame columns.
}

