% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/client.R
\docType{data}
\name{redshift.schema}
\alias{redshift.schema}
\title{Current database schema (to change, use redshift.connect())}
\format{\preformatted{ NULL
}}
\usage{
redshift.schema
}
\description{
Current database schema (to change, use redshift.connect())
}
\keyword{datasets}

