% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/client.R
\name{redshift.select}
\alias{redshift.select}
\title{Select data from database}
\usage{
redshift.select(sql, ...)
}
\arguments{
\item{sql}{Query string, may contain placeholders ? for parameters}

\item{...}{Query parameters}
}
\value{
A dataframe with results
}
\description{
Select data from database
}

