% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/client.R
\name{redshift.tableExists}
\alias{redshift.tableExists}
\title{Verify that a table exists in database}
\usage{
redshift.tableExists(tableName)
}
\arguments{
\item{tableName}{Name of the table (without schema).}
}
\value{
TRUE if the table exists, FALSE otherwise.
}
\description{
Verify that a table exists in database
}

