% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/client.R
\name{redshift.update}
\alias{redshift.update}
\title{Update/Insert data to database}
\usage{
redshift.update(sql, ...)
}
\arguments{
\item{sql}{Query string, may contain placeholders ? for parameters}

\item{...}{Query parameters}
}
\value{
void
}
\description{
Update/Insert data to database
}

