% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/client.R
\docType{class}
\name{RedshiftDriver-class}
\alias{RedshiftDriver}
\alias{RedshiftDriver-class}
\title{Class to access Redshift database}
\description{
Class to access Redshift database
}
\section{Fields}{

\describe{
\item{\code{conn}}{Database connection (JDBCConnection)}

\item{\code{schema}}{Current database schema}
}}

